Adding attributions

+++ attribution/code/seed.py
@@ -51,4 +51,4 @@ def gather_laws():
     """ Load tensors for main Laws datafile. """
     X_train, y_train = torch.load("../data/laws/tensor_train.pt")
     X_test, y_test = torch.load("../data/laws/tensor_test.pt")
-    return X_train, X_test, y_train, y_test
+    return X_train.float(), X_test.float(), y_train.float(), y_test.float()
